:art: llm 요약 코드 utils로 이동, 예시 코드 삭제
