+++ app/utils/llm.py
@@ -1,0 +1,31 @@
+from langchain.prompts.chat import (
+    ChatPromptTemplate,
+    SystemMessagePromptTemplate,
+    HumanMessagePromptTemplate,
+)
+from langchain.chains import LLMChain
+from langchain_community.chat_models import ChatOllama
+from langchain_core.output_parsers import StrOutputParser
+from langchain_core.prompts import ChatPromptTemplate
+class LLM:
+    def __init__(self):
+
+        template = """
+        너는 긴 계시물을 요약해주는 AI야.
+        5줄로 요약해주면 돼.
+        
+        글:"""
+        system_message_prompt = SystemMessagePromptTemplate.from_template(template)
+        human_template = "{text}"
+        human_message_prompt = HumanMessagePromptTemplate.from_template(human_template)
+
+        chat_prompt = ChatPromptTemplate.from_messages([system_message_prompt, human_message_prompt])
+
+        llm = ChatOllama(
+            model="gemma2"
+        )  # assuming you have Ollama installed and have llama3 model pulled with `ollama pull llama3 `
+        self.chain = chat_prompt | llm
+    def call(self,content:str):
+        return self.chain.invoke({"text":content})
+
+
